requirements.txt
# Core Dependencies
requests>=2.28.0
python-dotenv>=1.0.0

# LLM Dependencies (adjust based on your Llama implementation)
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Web Framework (if needed for reports)
flask>=2.3.0
jinja2>=3.1.0

# Validation & Testing
pydantic>=2.0.0
pytest>=7.0.0
pytest-cov>=4.0.0

# Utilities
pyyaml>=6.0
colorlog>=6.7.0
watchfiles>=0.19.0
rich>=13.0.0
textual>=0.40.0
halo>=0.0.31
asciichartpy>=1.5.11

# Development Dependencies
black>=23.0.0
mypy>=1.0.0
pre-commit>=3.0.0

# API server for approvals
fastapi>=0.95.0
uvicorn>=0.22.0

